# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Royal Midnight luxury black car service API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/bookings:
    get:
      operationId: getUserBookings
      tags: [users]
      summary: Get a user's booking history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

  # ── SAVED ADDRESSES ────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: driverId
          in: query
          schema:
            type: ["integer", "null"]
        - name: bookingId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── SUPPORT TICKETS ────────────────────────────────────────────────────────
  /support:
    get:
      operationId: listTickets
      tags: [support]
      summary: List support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createTicket
      tags: [support]
      summary: Submit a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketBody"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/{id}:
    patch:
      operationId: updateTicket
      tags: [support]
      summary: Update ticket status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketBody"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"

  # ── NOTIFICATIONS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [passenger, driver, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [passenger, driver, admin]
      required: [name, email, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    # ── VEHICLE SCHEMAS ───────────────────────────────────────────────────────
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        passengerName:
          type: string
        passengerEmail:
          type: string
        passengerPhone:
          type: string
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        pickupAt:
          type: string
          format: date-time
        vehicleClass:
          type: string
          enum: [standard, business, first_class, suv, van]
        passengers:
          type: integer
        flightNumber:
          type: ["string", "null"]
        specialRequests:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        priceQuoted:
          type: number
        promoCode:
          type: ["string", "null"]
        discountAmount:
          type: ["number", "null"]
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, passengerName, passengerEmail, passengerPhone, pickupAddress, dropoffAddress, pickupAt, vehicleClass, passengers, status, priceQuoted, createdAt, updatedAt]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        driverId:
          type: integer
        userId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, bookingId, driverId, rating, createdAt]
    CreateReviewBody:
      type: object
      properties:
        bookingId:
          type: integer
        driverId:
          type: integer
        userId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required: [bookingId, driverId, rating]

    # ── SUPPORT SCHEMAS ───────────────────────────────────────────────────────
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        bookingId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, subject, message, status, priority, createdAt, updatedAt]
    CreateTicketBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        bookingId:
          type: ["integer", "null"]
      required: [name, email, subject, message, priority]
    UpdateTicketBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, urgent, null]

    # ── NOTIFICATION SCHEMAS ──────────────────────────────────────────────────
...